// src/interfaces/vehiculo.Interface.ts
export interface Vehiculo {
    marca: string;
    icono: string;
    linea: string;
    color: string;
    modelo: number;
    numeroSerie: string;
    numeroMotor: string;
    numeroChasis: string;
    numeroVin: string;
    cilindraje: number;
    tipoCarroceria: string;
    tipoCombustible: string;
    autoridadTransito: string;
    placa: string;
    fechaMatricula: string;
}